add PfsSpectra
PfsSpectra is a collection of spectra from a common source (e.g., a single
arm from a spectrograph; or the entire instrument). It is intended that
this will replace PfsArm, which needs to be generalised. I have tried to
keep the persisted format as similar to that of PfsArm as possible.