add PfsSpectrum
This is a redesign of PfsObject, with a new name because PfsObject is
going to be a data type rather than the base class of single-spectrum
data types. This includes more data than the old PfsObject (e.g., the
observations), but I've tried to keep the persisted format as similar
to PfsObject as possible.