DAMD-49: replaced expId with visit
It has been agreed as part of DAMD-49 that the term
'visit' should be used instead of 'exposure'.
Consequently 'expId' should be reverted to the original
term used for the identifier, the 'visit' number.

In the case of PsfConfig, this makes use of the special visit
number, visit0, that corresponds to the first visit for which
the flats/biases/darks for a given set of science visits
are valid. That class is updated such that 'expId' is replaced
with 'visit0'.