pfsArm: make datamodel and code correspond
The datamodel specified a binary table with the pfsDesignId and visit.
The implementation didn't have such a binary table, so added it; but the
name is better described as IDENTITY rather than CONFIG because it's not
the full configuration.

Fixed the order of the HDUs to correspond to the implementation.