PfsSpectrum: allow use of WavelengthArray for wavelengths
A WavelengthArray is a numpy ndarray that is specified by the min and
max wavelength and length: like running np.linspace, except it keeps track
of the original parameters, and uses them for persistence via WCS headers.
This allows a modest savings in space for the PfsSpectrum over a full array
of wavelength values.