Replace calculate_pfsVisitHash with calculatePfsVisitHash
The old function wasn't used anywhere, and didn't sort the visits (so a
permuted list produces a different hash).
calculatePfsVisitHash checks for duplicates and sorts the list, and is
used (in TargetData.calculateVisitHash).
The new name for the function forces everyone to use the correct version.

As part of these changes, TargetData.getIdentity will now set nVisit to
the number of unique visits, so we don't double count when we have red
and blue arms.