fix docstrings for DRP products
PfsArm, PfsMerged, PfsReference, PfsSingle, PfsObject inherit from base
classes that have methods that may refer to the base class in their
docstrings. We don't want the user of these classes to notice the base
class names, so modify the docstrings with our new class decorator.

Added a test to ensure this remains the case.