datamodel: replace pfsFiberTrace with pfsFiberProfiles
Because the FiberTrace contains the profile realised for a particular trace
position, it's not possible to realise the profile at a different position
(it's undersampled, so we can't just resample). In order to be able to
realise the trace at a different position, we need to store the profile,
and not the realised profile.