make arrays have consistent types
in order to call into pybind APIs without casting by the user.
This involves changing most array outputs to double
precision (wavelength in particular needs to be double, and
then making most things double is convenient), and casting all
arrays to consistent types on read so the user doesn't have to
(and to avoid byte swapping problems).