Ensured that previous version of pfsConfig/Design files are readable.
To avoid frequent regeneration of weekly test data, allow the
previous version of pfsConfig/Design files to be readable.
This is done by making the new keyword introduced, ARMS,
optional on read.

Note that this is a temporary measure, and the new keyword will
be updated to be mandatory after the test data have been generated.