add versions to persistence of all classes
and document the versions in a new top-level text file.

We use the DAMD_VER header keyword to write an integer
version number for (almost) all classes. The new
versions.txt file lists the values and any changes.

Added comments in the write methods reminding developers
that any changes to the format need to increment the
version number and document the changes.