DifferentialDetectorMap: suppress VerifyWarning on write
Otherwise, they can overwhelm the stdout.
Not sure what astropy is warning about: all the header
keywords that are longer than 8 chars have HIERARCH
prepended already, but astropy says it's adding it.
There seems to be a difference between dotted
namespaces and single names: the former don't produce
warnings, but the latter do. But we don't care about
what might be a bug in astropy, so just suppress the
warnings.