write pfsArm in counts with explicit normalisation
Having pfsArm in counts makes it easier for instrument development and
direct comparison with the image.

We add an explicit NORM HDU containing the normalisation (the extracted
quartz flux from the fiberProfile) and carry that along so that we won't
have to go looking for it later (and risk using the wrong one).

Incremented the datamodel version to 2 for pfsArm/pfsMerged due to this
change.