Add implementation of PfsFluxReference class
PfsFluxReference class was defined by DAMD-127 and DAMD-129,
and it was named "PfsReference" then. This commit not only implements it
but also changes its name to "PfsFluxReference" because the old name
has already signified another class. PfsFluxReference will eventually
obsolete the old class, but this objective has not been achived yet.
Many things would break if the signified were to change now.