PfsConfig: get pfsDesignId,visit0 from header
It was never a good idea to rely on the filename to provide information
we should have gotten from the contents of the file. In particular, the
pfsDesignId and visit0 now get written to the FITS header, and read from
there if available. There is a fallback mechanism to get those from the
filename (assuming the filename matches the template). I also added a
script that will check an existing file and write the appropriate header
keywords if they are missing (so the file can be renamed without losing the
ability to read it).