add MultipleDistortionsDetectorMap
MultipleDistortionsDetectorMap allows layering distortions on top of each
other. Distortions can be of different kinds (we currently have only
DoubleDistortion, which is a separate distortion field for the left and
right CCDs, but we are going to want to add empirical distortion fields
eventually, to account for the tree rings). Layering distortions allows
more flexibility than tweaking a single distortion field.