PfsTable: expand schema requirements and add HDU-based I/O
Now require a documentation string and default value for each column. This
is provided in a new dataclass, Column. The schema is now an iterable of
Column; the getSchemaDict method can provide the old version of the schema.

Added readHdu and writeHdu methods that implement the I/O for an open FITS
file. This will allow a table to be written as part of a larger FITS file.

These changes are in support of the new "reduction notes" facility.