refactor PfsFiberNorms
After some discussion, PfsFiberNorms has been redesigned. It now includes
the normalisation value for each pixel of each fiber. The wavelength is
also included, but more for convenience than a real need (because it's
available in the pfsArm). We also include a table of model parameters. We
expect that this model will evolve; it is currrently specified as a FITS
HDU, so the datamodel package doesn't need to worry about how to read it.
We need to record the rotation angle of the quartzes used to measure the
fiberNorms, so we can apply the correction from the screen to the sky.
Finally, we're including hashes of multiple fiberProfiles, since we will
write one fiberNorms per arm for all spectrographs; this will allow us to
enforce that a fiberNorms is constructed from the same fiberProfiles used
in extracting the pfsArm.