PfsTargetSpectra: handle case where spectra have different lengths
The usual case is that the spectra have the same length, since we use the
same wavelength sampling. If the wavelength arrays are all equal, we can
simply write one and use that for all.
If the spectra don't all have the same length, we need to use tables to
write the arrays rather than images.