Add a new TargetType SCIENCE_MASKED
Before a pfsConfig is sent to open-use PIs, it will be modified
so that the other PIs' targets will be replaced with nonsense numbers.
The 2D pipeline must ignore these masked targets when the PIs run it
for themselves. We expect the 2D pipeline will ignore targets that
are not "SCIENCE". We therefore introduce a new TargetType
SCIENCE_MASKED indicating "The target is redacted for privacy."